Finished air current object and functionality.
I opted for using two different settings for the water vapour and the player instead of keeping a list of all objects in contact with the current and updating all separately. It seems to be the best approach for us since we won't have many objects interacting with the air currents.